+++ Platform/OpenCanopy/OcBootstrap.c
@@ -104,12 +104,16 @@ OcShowMenuByOc (
   OcSetInitialCursorOffset ();
   mGuiContext.BootEntry   = NULL;
   mGuiContext.ReadyToBoot = FALSE;
+
   //
-  // Re-run intro animation on each entry into menu, to avoid stuck animation
-  // which happens otherwise, if a menu item which returns to the menu is
-  // selected before the animation ends.
+  // When enabled, re-run intro animation on each entry into menu, to avoid
+  // stuck animation which happens otherwise, if a menu item which returns to
+  // the menu is selected before the animation ends.
+  // Do not play intro animation for visually impaired users.
   //
-  mGuiContext.DoneIntroAnimation   = FALSE;
+  mGuiContext.UseMenuEaseIn = !BootContext->PickerContext->PickerAudioAssist
+                              && ((BootContext->PickerContext->PickerAttributes & OC_ATTR_REDUCE_MOTION) == 0);
+
   mGuiContext.HideAuxiliary        = BootContext->PickerContext->HideAuxiliary;
   mGuiContext.Refresh              = FALSE;
   mGuiContext.PickerContext        = BootContext->PickerContext;
@@ -121,12 +125,6 @@ OcShowMenuByOc (
   }
 
   mDrawContext.TimeOutSeconds = BootContext->PickerContext->TimeoutSeconds;
-  //
-  // Do not play intro animation for visually impaired users.
-  //
-  if (BootContext->PickerContext->PickerAudioAssist) {
-    mGuiContext.DoneIntroAnimation = TRUE;
-  }
 
   Status = BootPickerViewInitialize (
              &mDrawContext,
@@ -254,12 +252,8 @@ OcShowPasswordByOc (
 
   mDrawContext.TimeOutSeconds = 0;
 
-  //
-  // FIXME: Ideally we disable password fade-in animation here for visually
-  // impaired users, i.e. when Context->PickerAudioAssist is TRUE, as we
-  // already do for menu ease-in animation. Probably needs separate bool in
-  // addition to mGuiContext.DoneIntroAnimation.
-  //
+  mGuiContext.UsePasswordEaseIn = !Context->PickerAudioAssist
+                                  && ((Context->PickerAttributes & OC_ATTR_REDUCE_MOTION) == 0);
 
   Status = PasswordViewInitialize (
              &mDrawContext,