+++ Platform/OpenCanopy/Views/BootPicker.c
@@ -2028,7 +2028,7 @@ BootPickerViewInitialize (
 
   InitializeListHead (&mBootPickerLabelAnimation.Link);
 
-  if (!GuiContext->DoneIntroAnimation) {
+  if (GuiContext->UseMenuEaseIn) {
     InitBpAnimIntro (DrawContext);
     InsertHeadList (&DrawContext->Animations, &mBootPickerIntroAnimation.Link);
     //
@@ -2037,8 +2037,6 @@ BootPickerViewInitialize (
     mBootPickerContainer.Obj.Opacity       = 0;
     mBootPickerLeftScroll.Hdr.Obj.Opacity  = 0;
     mBootPickerRightScroll.Hdr.Obj.Opacity = 0;
-
-    GuiContext->DoneIntroAnimation = TRUE;
   } else {
     //
     // The late code assumes the scroll buttons are visible by default.
@@ -2051,7 +2049,9 @@ BootPickerViewInitialize (
     mCommonActionButtonsContainer.Obj.Opacity = 0xFF;
   }
 
-  if (DrawContext->TimeOutSeconds > 0) {
+  if (  !GuiContext->PickerContext->PickerAudioAssist
+     && (DrawContext->TimeOutSeconds > 0))
+  {
     STATIC GUI_ANIMATION  PickerAnim2;
     PickerAnim2.Context = NULL;
     PickerAnim2.Animate = InternalBootPickerAnimateTimeout;